fix: harden authoritative catalog full refresh

+++ src/Win7POS.Core/Online/OnlineSyncSupervisorContracts.cs
@@ -142,8 +142,8 @@ namespace Win7POS.Core.Online
             bool requestCatalogNow = false,
             int? nextPollAfterSeconds = null,
             bool terminal = false,
-            int catalogPagesProcessed = 0,
-            int catalogRowsApplied = 0,
+            long catalogPagesProcessed = 0,
+            long catalogRowsApplied = 0,
             bool catalogSaleSafe = false,
             PosRuntimeDiagnostic catalogDiagnostic = null)
         {
@@ -173,8 +173,8 @@ namespace Win7POS.Core.Online
 
         public bool AuthenticationDenied { get; }
         public bool CatalogHasMore { get; }
-        public int CatalogPagesProcessed { get; }
-        public int CatalogRowsApplied { get; }
+        public long CatalogPagesProcessed { get; }
+        public long CatalogRowsApplied { get; }
         public bool CatalogSaleSafe { get; }
         public PosRuntimeDiagnostic CatalogDiagnostic { get; }
         public string Code { get; }
